services/image-processor/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pillow>=10.2.0
pydantic>=2.10.0
python-multipart>=0.0.12
